simulation debug

+++ stim/Stimulator.py
@@ -9,8 +9,8 @@ class Stimulator(object):
         self.start        = start
         self.train        = train
         self.amplitude    = abs(amplitude)  # (uA/cm^2)
-        self._map_on      = np.zeros( self.shape, dtype=np.float32 )
-        self._map_off     = np.zeros( self.shape, dtype=np.float32 )
+        self._map_on      = np.zeros( self.shape, dtype=np.float64 )
+        self._map_off     = np.zeros( self.shape, dtype=np.float64 )
     
     def get_flag(self, t):
         if t >= self.start and t < self.start + self.train * self.interval and (t - self.start) % self.interval < self.duration:
